+++ packages/twopoint5d/src/stage/public-api.ts
@@ -1,6 +1,7 @@
 export * from './Canvas2DStage.js';
 export * from './ClearStage.js';
 export * from './fitIntoRectangle.js';
+export type * from './IPassProvider.js';
 export type * from './IProjection.js';
 export type * from './IRenderable.js';
 export type * from './IStage.js';
@@ -8,5 +9,6 @@ export type * from './IStageRendererHost.js';
 export * from './OrthographicProjection.js';
 export * from './ParallaxProjection.js';
 export * from './ProjectionPlane.js';
+export * from './RootRenderPipeline.js';
 export * from './Stage2D.js';
 export * from './StageRenderer.js';